Always restart job search from beginning.
When a job is started, site_update_on_run() modifies share usage,
which might change job order when the jobs are sorted. Until we
can get smarter about when this might happen, assume it always
happens and so we need to start our search at the beginning of
the job list every time.